Initial commit for liby4m
This project is currently in the following state:

The header parser seems to be working correctly.

I'm still unsure about the best interface for accessing the data of the
image at any co-ordinate, but I think I'll start with the one
outlined in the header file. None of that has been implemented
yet, and I would need more information on the byte structure of
each of the colour spaces before I could implement it. This is
the next step.

The build system is a bit of a joke. Extremely static to say the
least. Eventually I'd like to get a cmake system going so that
it would be more portable and so that I could get experience with
the cmake build system that I love so much. Right now it's just
a makefile which compiles the library into the bin folder.

The test directory includes a test program, and a makefile which
makes use of the '../bin' directory to get the library, so the
library must be built first.

The parse module is supposed to be a woker function for the library,
but I have the funniest feeling that those functions are being
exported by the shared object, so I might have to look into that.