Implement reading of the frame header
Updating test to go through that path as well.

At this point, frame parameters will over-write header ones,
but seeing as the header will never be re-visited, this could
be viewed as the correct behaviour. It isn't clear which types
of parameters are allowed in the FRAME definition, or if any are
any that cannot be there.
This will need to be changed later when trying to write the file
back to the disk, seeing as we are throwing information away.
Warnings should be put in for this case later on.