Changing the frame reading behaviour
Previously, I was only holding the current frame in memory.
This was a good idea, but doesn't allow for persistance of
changes to that memory made by the application, and also
doesn't allow the frames to be written to a file very easily.

In order to eventually allow writing the changes to a file to disk,
I've changed it so that the library still only reads the data when
asked by the application, but then keeps all previous frame data
in memory. This will allow for the possibility of writing all of
the data to a new file on disk if the application so wishes.

The frame data is now held in a linked list structure for easy
storage.

This is obviously a trade-off between simplicity and memory. If
this library is to be able to deal with huge files, the logic for
writing the file back to disk will need to become more complex,
and the data from the file could be made to be read in frame by frame.