All the plumbing for writing the file back
Compiled and tested. Test commit to come.

So now the interface can be used to write the changes that are made
to the pointers to the data back into a file!
Yay.
I don't know what else to say about it other than the interface could
probably use some work. It also has the glaring issue that per frame
options aren't respected. I haven't seen any examples of that yet
though, so maybe it's no big deal.