Adding a license file
I should probably add this to the top of all the source code files.

This choice was not taken lightly. I went with MIT over GPL because
I don't envision this code base becoming very large, and I want people
to be able to use it without worrying too much.
I did not license it MIT because I wasn't worried about getting back
contributions. I'd still really appreciate pull requests for improvements. :D

Above all I hope this is useful to somebody, which is the other reason
for choosing MIT.