Adding CMake build system
As far as I can tell this kind of works, but I'm not really
an expert. It allows me to install the library in standard system
folders, and then link to them from the test (which is admittedly
kind of fragile what with hardcoding /usr/local/lib).

Currently the test cmake file is separate from the library test.
This kind of makes sense seeing as the test directory just houses
development testing grounds as opposed to any real testing framework.